refactor: extract shared shell/git test helpers into TestSupport target
Three test files (WorktreeLifecycleTests, WorktreeAdoptTests,
ArchivedReviveBranchRenameTests) each duplicated bash/git/repo
fixtures. Consolidate into a new TestSupport target so future tests
can reuse them.

- shell(_ command: String, at: URL) — hermetic git env
- createTestRepo() / createTestRepoResolvingSymlinks()
- makeRepoWithExternalWorktree(branch:folder:)
- makeTestRepo(db:tempDir:repoDir:) — overrides worktreeRoot
