+++ Package.swift
@@ -84,6 +84,14 @@ let package = Package(
                 .unsafeFlags(["-no-whole-module-optimization"], .when(configuration: .debug)),
             ]
         ),
+        .target(
+            name: "TestSupport",
+            dependencies: [
+                "TBDDaemonLib",
+                "TBDShared",
+            ],
+            path: "Tests/TestSupport"
+        ),
         .testTarget(
             name: "TBDSharedTests",
             dependencies: ["TBDShared"]
@@ -93,6 +101,7 @@ let package = Package(
             dependencies: [
                 "TBDDaemonLib",
                 "TBDShared",
+                "TestSupport",
                 .product(name: "GRDB", package: "GRDB.swift"),
             ]
         ),
