+++ Tests/TBDDaemonTests/WorktreeLifecycleTests.swift
@@ -1,65 +1,8 @@
 import Foundation
+import TestSupport
 import Testing
 @testable import TBDDaemonLib
 @testable import TBDShared
-
-/// Helper to run shell commands in tests.
-private func shell(_ command: String, at dir: URL) async throws {
-    let process = Process()
-    process.executableURL = URL(fileURLWithPath: "/bin/bash")
-    process.arguments = ["-c", command]
-    process.currentDirectoryURL = dir
-    process.environment = [
-        "PATH": "/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin",
-        "HOME": NSHomeDirectory(),
-        "GIT_CONFIG_NOSYSTEM": "1",
-        "GIT_CONFIG_GLOBAL": "/dev/null",
-        "GIT_AUTHOR_NAME": "Test",
-        "GIT_AUTHOR_EMAIL": "test@test.com",
-        "GIT_COMMITTER_NAME": "Test",
-        "GIT_COMMITTER_EMAIL": "test@test.com",
-    ]
-    let pipe = Pipe()
-    process.standardOutput = pipe
-    process.standardError = pipe
-    try process.run()
-    process.waitUntilExit()
-    guard process.terminationStatus == 0 else {
-        let data = pipe.fileHandleForReading.readDataToEndOfFile()
-        let output = String(data: data, encoding: .utf8) ?? ""
-        throw NSError(
-            domain: "shell",
-            code: Int(process.terminationStatus),
-            userInfo: [NSLocalizedDescriptionKey: "Command failed: \(command)\n\(output)"]
-        )
-    }
-}
-
-/// Creates a temporary git repo for testing, returning the temp dir and repo dir URLs.
-/// The caller is responsible for cleaning up the temp dir.
-private func createTestRepo() async throws -> (tempDir: URL, repoDir: URL) {
-    let tempDir = FileManager.default.temporaryDirectory
-        .appendingPathComponent("tbd-test-\(UUID().uuidString)")
-    let repoDir = tempDir.appendingPathComponent("repo")
-    try FileManager.default.createDirectory(at: repoDir, withIntermediateDirectories: true)
-    try await shell("git init -b main && git commit --allow-empty -m 'init'", at: repoDir)
-    return (tempDir: tempDir, repoDir: repoDir)
-}
-
-/// Creates a test repo row in the DB and overrides its `worktreeRoot` to a
-/// `.tbd/worktrees/` subdirectory of the test temp dir, so the canonical
-/// layout doesn't leak into the user's real `~/.tbd/worktrees/`. Returns the
-/// re-fetched repo with the override applied.
-private func makeTestRepo(
-    db: TBDDatabase, tempDir: URL, repoDir: URL
-) async throws -> Repo {
-    let repo = try await db.repos.create(
-        path: repoDir.path, displayName: "test", defaultBranch: "main"
-    )
-    let override = tempDir.appendingPathComponent(".tbd/worktrees").path
-    try await db.repos.updateWorktreeRoot(id: repo.id, path: override)
-    return try await db.repos.get(id: repo.id)!
-}
 
 @Test func testCreateWorktree() async throws {
     let (tempDir, repoDir) = try await createTestRepo()
@@ -567,29 +510,6 @@ private func makeTestRepo(
 }
 
 // MARK: - Reconcile Tests
-
-/// Like `createTestRepo()` but resolves symlinks in the returned paths so the
-/// path stored in the DB matches what `git worktree list` reports.
-///
-/// On macOS, `FileManager.default.temporaryDirectory` returns `/var/folders/…`
-/// which is a symlink to `/private/var/folders/…`. `URL.resolvingSymlinksInPath()`
-/// does NOT resolve this particular symlink, but the C `realpath()` function does.
-/// Git resolves the real path when recording worktree entries, so DB paths must also
-/// use the real path for reconcile path-matching to succeed.
-private func createTestRepoResolvingSymlinks() async throws -> (tempDir: URL, repoDir: URL) {
-    let (rawTempDir, _) = try await createTestRepo()
-    // Use C realpath() to fully resolve all symlinks (URL.resolvingSymlinksInPath
-    // does not resolve the /var → /private/var symlink on macOS).
-    let resolved: URL
-    if let cReal = realpath(rawTempDir.path, nil) {
-        resolved = URL(fileURLWithPath: String(cString: cReal))
-        free(cReal)
-    } else {
-        resolved = rawTempDir
-    }
-    let repoDir = resolved.appendingPathComponent("repo")
-    return (tempDir: resolved, repoDir: repoDir)
-}
 
 /// Alive server + alive windows: no terminals are deleted, window IDs unchanged.
 /// (dryRun makes serverExists → true and windowExists → true, so this exercises
